Async op samples usage (#465)
This adds a "samples" directory with sample usage of async_op. It shows how async_op may be used by various compositions of layered modules, including when a module calls another module with asynchronous retries or when a module must cancel all pending async_ops that are in-flight when it is closed.

The samples include some basic test coverage (though they do not have any "chaos" testing) to validate.

For now, this serves as a sample set of "design patterns", but it may be possible to extract some of this into macros or other modules for re-use.